+++ src/main/java/module-info.java
@@ -1,6 +1,5 @@
 module org.app {
     requires javafx.fxml;
-    requires javafx.web;
     requires static lombok;
     requires jakarta.annotation;
     requires atlantafx.base;
@@ -10,7 +9,10 @@ module org.app {
     requires org.kordamp.ikonli.javafx;
     requires org.kordamp.ikonli.core;
     requires org.kordamp.ikonli.material2;
+    requires org.kordamp.ikonli.lineawesome;
     requires org.slf4j;
+    requires com.gluonhq.richtextarea;
+    requires com.gluonhq.emoji;
 
     exports org.app;
     exports org.controller;