+++ src/main/java/org/controller/JournalController.java
@@ -1,6 +1,7 @@
 package org.controller;
 
 import atlantafx.base.theme.Styles;
+import com.gluonhq.richtextarea.model.Document;
 import javafx.collections.ListChangeListener;
 import javafx.collections.transformation.FilteredList;
 import javafx.event.ActionEvent;
@@ -11,6 +12,7 @@ import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.TreeItemPropertyValueFactory;
 import javafx.scene.layout.Pane;
+import javafx.scene.layout.VBox;
 import javafx.util.Callback;
 import org.context.ControllerRegistry;
 import org.context.GlobalContext;
@@ -40,8 +42,6 @@ public class JournalController extends Pane implements Initializable, SaveHandle
 
     @FXML
     public Label symbolLabel;
-    @FXML
-    public TextArea textArea;
 
     @FXML
     public Button saveBtn;
@@ -49,6 +49,9 @@ public class JournalController extends Pane implements Initializable, SaveHandle
     public Button deleteSymbol;
     @FXML
     public Button deleteDay;
+    @FXML
+    public VBox editor;
+    private RichTextEditorController editorController;
 
     FilteredList<Journal> journalEntries = GlobalContext.getJournals().getFiltered();
     private Node addJournalEntry;
@@ -59,10 +62,19 @@ public class JournalController extends Pane implements Initializable, SaveHandle
     public void initialize(URL location, ResourceBundle resources) {
         ///Defaults
         ControllerRegistry.register(JournalController.class, this);
+        try {
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("/org/app/fxml/RichTextEditor.fxml"));
+            Node editorNode = loader.load();
+            editorController = loader.getController();
+            editor.getChildren().add(editorNode);
+        } catch (IOException exception) {
+            throw new RuntimeException(exception);
+        }
         saveBtn.getStyleClass().add(Styles.ACCENT);
         deleteSymbol.setDisable(true);
         deleteDay.setDisable(true);
         saveBtn.setDisable(true);
+//        editorController.getEditor().setDisable(true);
         saveBtn.setOnAction(event -> save());
 
         ///Populate list
@@ -83,6 +95,7 @@ public class JournalController extends Pane implements Initializable, SaveHandle
             if (newItem != null) {
                 selectedSymbol = (Journal) newItem.getValue();
                 resetFormWithSelectedValue(selectedSymbol);
+//                editorController.getEditor().editableProperty().setValue(selectedSymbol.getDate() == null);
                 deleteDay.setDisable(selectedSymbol.getDate() == null);
                 deleteSymbol.setDisable(selectedSymbol.getSymbol() == null);
                 saveBtn.setDisable(selectedSymbol.getSymbol() == null);
@@ -131,7 +144,7 @@ public class JournalController extends Pane implements Initializable, SaveHandle
                             new TreeItem<>(new Journal(null, date, null, null));
                     for (Journal j : list) {
                         dateNode.getChildren().add(
-                                new TreeItem<>(new Journal(j.getId(), null, j.getSymbol(), j.getText()))
+                                new TreeItem<>(new Journal(j.getId(), null, j.getSymbol(), j.getDocument()))
                         );
                     }
 
@@ -142,7 +155,7 @@ public class JournalController extends Pane implements Initializable, SaveHandle
     @FXML
     @Override
     public void save() {
-        selectedSymbol.setText(textArea.getText());
+        selectedSymbol.setDocument(editorController.getDocument());
         DbManager db = new DbManager();
         try {
             db.setBdConnection();
@@ -161,11 +174,11 @@ public class JournalController extends Pane implements Initializable, SaveHandle
 
     void resetFormWithSelectedValue() {
         selectedSymbol = null;
-        textArea.clear();
+        editorController.setDocument(new Document());
     }
 
     void resetFormWithSelectedValue(Journal journal) {
-        textArea.setText(journal.getText());
+        editorController.setDocument(journal.getDocument());
     }
 
     @FXML